Deleteing links and nodes
